+++ packages/coverage-aggregator/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.4](https://github.com/synapsecns/sanguine/compare/@synapsecns/coverage-aggregator@1.0.3...@synapsecns/coverage-aggregator@1.0.4) (2024-04-04)
+
+**Note:** Version bump only for package @synapsecns/coverage-aggregator
+
+
+
+
+
+## 1.0.3 (2024-04-04)
+
+**Note:** Version bump only for package @synapsecns/coverage-aggregator
+
+
+
+
+
 ## 1.0.1 (2023-02-15)
 
 **Note:** Version bump only for package @synapsecns/coverage-aggregator